# config/openapi.yml
openapi: '3.0.2'
info:
  title: Proton API
  description: Backend API example API declaration
  version: '1.0'
servers:
  - url: /v1
paths:
...
  /users:
    post:
      description: Adds a User to the list
      operationId: postUser
      responses:
        201:
          $ref: '#/components/responses/201'
        409:
          $ref: '#/components/responses/400'
  /users/activations:
    put:
      description: Updates a User activation state
      operationId: putUserActivation
      parameters:
        - in: query
          name: uat
          required: true
          schema:
            type: string
          description: User Token
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
  /users/credentials:
    get:
      description: Gets a User session by credentials
      operationId: getUserSessionByCredentials
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
  /users/{id}:
    get:
      description: Get a User by Id
      operationId: getUserById
      security:
        - bearerAuth: []
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
    put:
      description: Updates a User at the list
      operationId: putUser
      security:
        - bearerAuth: []
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
    delete:
      description: Deletes a User from the list
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        204:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    '200':
      description: Successful call
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/200'
    '201':
      description: Successful added call
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/201'
...
    '400':
      description: Bad request. Check your parameters or payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/400'
...
    '404':
      description: The resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/404'
...
  schemas:
    200:
      title: Successful processed response
      type: object
    201:
      title: Successful added response
      type: object
...
    400:
      title: Request error response
      type: object
...
    404:
      title: Not found response
      type: object
...